// src/app/models/pessoa-crud.model.ts
export class PessoaImportacao {
    id: number = 0;
    nome: string = '';
    cpf: string = '';
    dataNascimento: Date = '' as unknown as Date;
    situacao: string = '';
    dataInscricao:  Date = '' as unknown as Date;
    digito: string = '';
    anoObito: string = '';
    excel_Status: string = '';
    excel_Data_Cap: Date = '' as unknown as Date;
    excel_Hora_Cap: Date = '' as unknown as Date;
    excel_IdNum: string = '';
    excel_Controle: string = '';
    excel_Erro: string = '';
    isDuplicate = true;
    isValid = true;
}